rag-service/requirements.txt
# Using >= so pip can pick pre-built wheels for Python 3.13 on Windows.
# The == pins in the original guide predated Python 3.13 and have no wheels.
sentence-transformers>=2.7.0
chromadb>=0.5.3
openai>=1.30.0
anthropic>=0.28.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pandas>=2.2.2
python-dotenv>=1.0.0
pydantic>=2.7.0
